implemented cache for universal search and seperated out service file for account from the controller
