+++ Projects/kz-srikar/PaginationBackend/src/main/java/com/example/pagination/PaginationBackendApplication.java
@@ -3,8 +3,8 @@ package com.example.pagination;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.Ordered;
 import org.springframework.web.cors.CorsConfiguration;
@@ -13,6 +13,7 @@ import org.springframework.web.filter.CorsFilter;
 import java.util.Collections;
 
 @SpringBootApplication
+@EnableCaching
 public class PaginationBackendApplication {
 
 	public static void main(String[] args) {
